+++ client/src/components/Ticketing/ticketingmanager/Season/components/EventCard.tsx
@@ -1,4 +1,4 @@
-import React, {useState} from 'react';
+import React from 'react';
 import {getImageDefault} from '../../../../../utils/imageURLValidation';
 import {useNavigate} from 'react-router';
 import {EventImage} from '../../../../../utils/imageURLValidation';
@@ -59,7 +59,7 @@ const EventCard = (props: EventCardProps) => {
       ) : (
         <article className='flex flex-col w-48 my-2'>
           <button
-            onClick={() => navigate(`/ticketing/showings/${Number(eventId)}`)}
+            onClick={() => navigate(`/ticketing/events/${Number(eventId)}`)}
             data-testid='event-page'
             className='bg-blue-500 hover:bg-blue-700 disabled:bg-gray-500 text-white py-2 px-3 rounded-xl mb-2'
             disabled={isFormEditing || isAddEventActive}
